Reorganize project access classes
