+++ web/modules/projects/projects/projects.module
@@ -7,7 +7,7 @@
 
 use Drupal\projects\Plugin\Field\UserIsManagerFieldItemList;
 use Drupal\projects\Plugin\Field\UserIsParticipantFieldItemList;
-use Drupal\projects\ProjectEntityAccess;
+use Drupal\projects\Access\ProjectEntityAccess;
 use Drupal\youvo\ComputedFieldStorageDefinition;
 use Drupal\Core\Access\AccessResult;
 use Drupal\Core\Field\FieldDefinitionInterface;
@@ -16,7 +16,7 @@ use Drupal\Core\Session\AccountInterface;
 use Drupal\multivalue_form_element\Element\MultiValue;
 use Drupal\projects\Entity\Project;
 use Drupal\projects\Plugin\Field\UserIsApplicantFieldItemList;
-use Drupal\projects\ProjectFieldAccess;
+use Drupal\projects\Access\ProjectFieldAccess;
 use Drupal\user\UserInterface;
 use Drupal\Core\Entity\EntityTypeInterface;
 use Drupal\Core\Field\BaseFieldDefinition;
