+++ web/modules/projects/projects/src/Access/ProjectActionAccess.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Drupal\projects;
+namespace Drupal\projects\Access;
 
 use Drupal\Core\Access\AccessResult;
 use Drupal\Core\Controller\ControllerBase;
 use Drupal\Core\Session\AccountInterface;
-use Drupal\projects\Entity\Project;
+use Drupal\projects\ProjectInterface;
 use Drupal\user_types\Utility\Profile;
 
 /**
@@ -18,14 +18,14 @@ class ProjectActionAccess extends ControllerBase {
    *
    * @param \Drupal\Core\Session\AccountInterface $account
    *   The account.
-   * @param \Drupal\projects\Entity\Project|null $project
+   * @param \Drupal\projects\ProjectInterface|null $project
    *   The project.
    *
    * @return \Drupal\Core\Access\AccessResultInterface
    *   The access results.
    */
   public function accessProjectApply(AccountInterface $account, ProjectInterface $project = NULL) {
-    if ($project instanceof Project) {
+    if ($project instanceof ProjectInterface) {
       return AccessResult::allowedIf(
         Profile::isCreative($account) &&
         $project->lifecycle()->isOpen()
@@ -39,14 +39,14 @@ class ProjectActionAccess extends ControllerBase {
    *
    * @param \Drupal\Core\Session\AccountInterface $account
    *   The account.
-   * @param \Drupal\projects\Entity\Project|null $project
+   * @param \Drupal\projects\ProjectInterface|null $project
    *   The project.
    *
    * @return \Drupal\Core\Access\AccessResultInterface
    *   The access results.
    */
   public function accessProjectNotify(AccountInterface $account, ProjectInterface $project = NULL) {
-    if ($project instanceof Project) {
+    if ($project instanceof ProjectInterface) {
       return AccessResult::allowedIf(
         $project->lifecycle()->isDraft() &&
         $project->isAuthorOrManager($account)
