+++ web/modules/projects/projects/src/Access/ProjectTransitionAccess.php
@@ -6,7 +6,6 @@ use Drupal\Core\Access\AccessResult;
 use Drupal\Core\Controller\ControllerBase;
 use Drupal\Core\Session\AccountInterface;
 use Drupal\lifecycle\Permissions;
-use Drupal\projects\Entity\Project;
 
 /**
  * Access controller for transition forms.
@@ -18,7 +17,7 @@ class ProjectTransitionAccess extends ControllerBase {
    *
    * @param \Drupal\Core\Session\AccountInterface $account
    *   The account.
-   * @param \Drupal\projects\Entity\Project|null $project
+   * @param \Drupal\projects\ProjectInterface|null $project
    *   The project.
    * @param string $transition
    *   The requested transition. Defaults in route definition.
@@ -27,7 +26,7 @@ class ProjectTransitionAccess extends ControllerBase {
    *   The access results.
    */
   public function accessProjectTransition(AccountInterface $account, ProjectInterface $project = NULL, string $transition = '') {
-    if ($project instanceof Project && !empty($transition)) {
+    if ($project instanceof ProjectInterface && !empty($transition)) {
       return AccessResult::allowedIf(
         Permissions::useTransition($account, 'project_lifecycle', $transition) &&
         $project->lifecycle()->canTransition($transition)
